backend/pyproject.toml
[tool.poetry]
name = "clausecompass-backend"
version = "0.1.0"
description = "ClauseCompass - AI-powered legal document analysis backend"
authors = ["ClauseCompass Team <team@clausecompass.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.116.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
httpx = "^0.28.1"
textstat = "^0.7.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
google-cloud-documentai = "^3.6.0"
google-genai = "1.33.0"
google-cloud-firestore = "^2.13.1"
google-cloud-storage = "^2.12.0"
google-cloud-dlp = "^3.14.0"
google-cloud-pubsub = "^2.31.1"
google-cloud-bigquery = "^3.13.0"
google-cloud-translate = "^3.15.0"
google-auth = "^2.25.2"
pandas = "^2.1.4"
numpy = "2.3.2"
scikit-learn = "^1.3.2"
regex = "2025.9.1"
PyPDF2 = "^3.0.1"
pdfminer-six = "^20231228"
google-generativeai = "^0.8.5"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "textstat.*",
    "PyPDF2.*",
    "pdfminer.*",
]
ignore_missing_imports = true